+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Support `Sdk.SuppressInstrumentation`/`SuppressInstrumentationScope`.
+  ([#3372](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3372))
+
 ## 1.14.0-beta.1
 
 Released 2025-Nov-13