+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated OpenTelemetry core component version(s) to `1.15.1`.
+  ([#4020](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4020))
+  
 * Commands that fail due to a connection timeout or connection failure
   now correctly set the trace `Activity` status to `Error`.
   ([#4019](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4019))