+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -5,7 +5,7 @@
 * Updated OpenTelemetry core component version(s) to `1.15.3`.
   ([#4166](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4166))
 
-* Fixed `db.query.text` emission to respect `SetVerboseDatabaseStatements` when
+* Fixed `db.query.text` not respecting `SetVerboseDatabaseStatements` when
   the new database semantic conventions are enabled.
   ([#4245](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4245))